refactor: extract FulfillmentCard into its own file
